Make changes to styles and file structures
