+++ src/index.html
@@ -14,7 +14,7 @@
     <div class="header--logo">
         <img src="https://t3.ftcdn.net/jpg/04/51/58/76/240_F_451587634_iVJKobsbjLpbtBY0HTkZ2tzgUVGHfzEU.jpg" alt="TV logo">
     </div>
-    <h3>Shows(<span class="home--counter">0</span>)</h3>
+    <h3> <span class="title">NELAK SHOWS</span> -  <span>Shows</span>(<span class="home--counter">0</span>)</h3>
   </header>
   <main>
     <section class="shows--container">
